refactor: remove unused and duplicate code
