+++ demo/dashboard/Dockerfile
@@ -4,4 +4,5 @@ COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 COPY app.py .
 COPY scenarios/ scenarios/
+COPY static/ static/
 CMD ["python", "app.py"]
