+++ src/shmem.rs
@@ -5,7 +5,7 @@
 
 use pgrx::prelude::*;
 use pgrx::{PGRXSharedMemory, PgAtomic, PgLwLock, pg_shmem_init};
-use std::sync::atomic::{AtomicU32, AtomicU64};
+use std::sync::atomic::AtomicU64;
 
 /// C2-1 / DEF-4: Maximum number of pgt_ids the invalidation ring buffer can hold.
 ///
@@ -270,15 +270,6 @@ pub static DAG_REBUILD_SIGNAL: PgAtomic<AtomicU64> =
 // SAFETY: PgAtomic::new requires a static CStr name.
 pub static CACHE_GENERATION: PgAtomic<AtomicU64> =
     unsafe { PgAtomic::new(c"pg_trickle_cache_gen") };
-
-/// Cluster-wide counter of active dynamic refresh workers.
-///
-/// Coordinators increment this before spawning a worker and workers decrement
-/// on exit. Used together with the `max_dynamic_refresh_workers` GUC to
-/// enforce a cluster-wide worker budget.
-// SAFETY: PgAtomic::new requires a static CStr name.
-pub static ACTIVE_REFRESH_WORKERS: PgAtomic<AtomicU32> =
-    unsafe { PgAtomic::new(c"pg_trickle_active_workers") };
 
 /// A reserved or live dynamic refresh worker slot.
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
@@ -1485,7 +1476,6 @@ pub fn init_shared_memory() {
     pg_shmem_init!(TICK_WATERMARK_STATE);
     pg_shmem_init!(DAG_REBUILD_SIGNAL);
     pg_shmem_init!(CACHE_GENERATION);
-    pg_shmem_init!(ACTIVE_REFRESH_WORKERS);
     pg_shmem_init!(WORKER_SLOT_TABLE);
     pg_shmem_init!(INVALIDATION_RING_OVERFLOWS);
     pg_shmem_init!(RECONCILE_EPOCH);
@@ -1949,75 +1939,8 @@ pub fn drain_status() -> Option<bool> {
     }
 }
 
-// ── Worker token management (Phase 2: parallel refresh) ───────────────────
-
-fn try_increment_bounded_counter(atomic: &AtomicU32, max_value: u32) -> bool {
-    loop {
-        let current = atomic.load(std::sync::atomic::Ordering::Acquire);
-        if current >= max_value {
-            return false;
-        }
-
-        match atomic.compare_exchange_weak(
-            current,
-            current + 1,
-            std::sync::atomic::Ordering::AcqRel,
-            std::sync::atomic::Ordering::Relaxed,
-        ) {
-            Ok(_) => return true,
-            Err(_) => continue,
-        }
-    }
-}
-
-fn saturating_decrement_counter(atomic: &AtomicU32) {
-    loop {
-        let current = atomic.load(std::sync::atomic::Ordering::Acquire);
-        let next = current.saturating_sub(1);
-        if next == current {
-            return;
-        }
-
-        match atomic.compare_exchange_weak(
-            current,
-            next,
-            std::sync::atomic::Ordering::AcqRel,
-            std::sync::atomic::Ordering::Relaxed,
-        ) {
-            Ok(_) => return,
-            Err(_) => continue,
-        }
-    }
-}
-
 fn increment_epoch(atomic: &AtomicU64) {
     atomic.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// Try to acquire a cluster-wide refresh worker token.
-///
-/// Returns `true` if a token was acquired (active count incremented),
-/// `false` if the budget is exhausted.
-///
-/// The caller **must** call [`release_worker_token`] when the worker
-/// finishes, regardless of success or failure.
-pub fn try_acquire_worker_token(max_workers: u32) -> bool {
-    if !is_shmem_available() {
-        return false;
-    }
-
-    try_increment_bounded_counter(ACTIVE_REFRESH_WORKERS.get(), max_workers)
-}
-
-/// Release a cluster-wide refresh worker token (decrement active count).
-///
-/// Called by workers on exit and by reconciliation logic for leaked tokens.
-pub fn release_worker_token() {
-    if !is_shmem_available() {
-        return;
-    }
-
-    saturating_decrement_counter(ACTIVE_REFRESH_WORKERS.get());
 }
 
 /// Read the current number of active dynamic refresh workers.
@@ -2064,18 +1987,6 @@ pub fn citus_worker_failure_total() -> u64 {
     CITUS_WORKER_FAILURE_TOTAL
         .get()
         .load(std::sync::atomic::Ordering::Relaxed)
-}
-
-/// Force-set the active worker count to a specific value.
-///
-/// Used by reconciliation logic to correct the count after a crash.
-pub fn set_active_worker_count(count: u32) {
-    if !is_shmem_available() {
-        return;
-    }
-    ACTIVE_REFRESH_WORKERS
-        .get()
-        .store(count, std::sync::atomic::Ordering::Release);
 }
 
 /// Bump the reconcile epoch (signals that token count was corrected).
@@ -2521,47 +2432,8 @@ static SHMEM_INITIALIZED: std::sync::atomic::AtomicBool = std::sync::atomic::Ato
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        PgTrickleSharedState, drain_ring, increment_epoch, push_invalidation,
-        saturating_decrement_counter, try_increment_bounded_counter,
-    };
-    use std::sync::atomic::{AtomicU32, AtomicU64, Ordering};
-
-    #[test]
-    fn test_try_increment_bounded_counter_increments_until_cap() {
-        let counter = AtomicU32::new(0);
-
-        assert!(try_increment_bounded_counter(&counter, 2));
-        assert_eq!(counter.load(Ordering::Relaxed), 1);
-
-        assert!(try_increment_bounded_counter(&counter, 2));
-        assert_eq!(counter.load(Ordering::Relaxed), 2);
-
-        assert!(!try_increment_bounded_counter(&counter, 2));
-        assert_eq!(counter.load(Ordering::Relaxed), 2);
-    }
-
-    #[test]
-    fn test_try_increment_bounded_counter_rejects_zero_budget() {
-        let counter = AtomicU32::new(0);
-
-        assert!(!try_increment_bounded_counter(&counter, 0));
-        assert_eq!(counter.load(Ordering::Relaxed), 0);
-    }
-
-    #[test]
-    fn test_saturating_decrement_counter_stops_at_zero() {
-        let counter = AtomicU32::new(2);
-
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 1);
-
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 0);
-
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 0);
-    }
+    use super::{PgTrickleSharedState, drain_ring, increment_epoch, push_invalidation};
+    use std::sync::atomic::{AtomicU64, Ordering};
 
     #[test]
     fn test_increment_epoch_advances_counter() {
@@ -2571,45 +2443,6 @@ mod tests {
         increment_epoch(&epoch);
 
         assert_eq!(epoch.load(Ordering::Relaxed), 43);
-    }
-
-    // ── P3: acquire/release token cycle and invariants ─────────────────────
-
-    #[test]
-    fn test_token_acquire_and_release_cycle() {
-        let counter = AtomicU32::new(0);
-
-        // Acquire two tokens up to budget of 3
-        assert!(try_increment_bounded_counter(&counter, 3));
-        assert!(try_increment_bounded_counter(&counter, 3));
-        assert_eq!(counter.load(Ordering::Relaxed), 2);
-
-        // Release one
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 1);
-
-        // Can acquire again
-        assert!(try_increment_bounded_counter(&counter, 3));
-        assert_eq!(counter.load(Ordering::Relaxed), 2);
-
-        // Release both
-        saturating_decrement_counter(&counter);
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 0);
-    }
-
-    #[test]
-    fn test_counter_does_not_go_below_zero_on_over_release() {
-        let counter = AtomicU32::new(1);
-
-        // Release once (valid)
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 0);
-
-        // Extra releases are no-ops
-        saturating_decrement_counter(&counter);
-        saturating_decrement_counter(&counter);
-        assert_eq!(counter.load(Ordering::Relaxed), 0);
     }
 
     #[test]
@@ -2623,21 +2456,6 @@ mod tests {
             assert!(current > last);
             last = current;
         }
-    }
-
-    #[test]
-    fn test_bounded_counter_budget_one_acts_as_mutex() {
-        let counter = AtomicU32::new(0);
-
-        // Acquire the single token
-        assert!(try_increment_bounded_counter(&counter, 1));
-        // Second acquire must be rejected
-        assert!(!try_increment_bounded_counter(&counter, 1));
-
-        // Release
-        saturating_decrement_counter(&counter);
-        // Now acquirable again
-        assert!(try_increment_bounded_counter(&counter, 1));
     }
 
     // ── C2-1: Invalidation ring buffer tests ──────────────────────────────
